+++ 2019/dag5/Program.cs
@@ -3,6 +3,7 @@ using System.IO;
 using AdventOfCode2019;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 
 namespace dag5
 {
@@ -16,7 +17,7 @@ namespace dag5
         static void Star(int input, int star)
         {
             var intcoder = new Intcoder(File.ReadAllText("input.txt"), OutputMode.OutputAndRunToEnd);
-            intcoder.Exec(new Queue<int>(new int[]{input}));
+            intcoder.Exec(new Queue<BigInteger>(new BigInteger[]{input}));
             Console.WriteLine($"Star {star}: {intcoder.Outputs.Last()}");
         }
     }
