+++ 2019/dag7/Program.cs
@@ -3,7 +3,7 @@ using AdventOfCode2019;
 using System.IO;
 using System.Collections.Generic;
 using System.Linq;
-using AdventofCode.Utils;
+using System.Numerics;
 
 namespace dag7
 {
@@ -18,14 +18,14 @@ namespace dag7
         static void Star2()
         {
             string input = File.ReadAllText("input.txt");
-            var thrusterSignals = new List<int>();
+            var thrusterSignals = new List<BigInteger>();
             var settings = Utils.GetPermutations(new List<int>(){5,6,7,8,9}, 5).Select(t => new List<int>(t));
             
             foreach (var parameters in settings)
             {
                 var intcoders = Enumerable.Range(0, 5).Select( t => new Intcoder(input, OutputMode.ReturnAndHalt)).ToArray();
                 int activRunning = 0;
-                var inputQueue = new Queue<int>(new int[]{parameters[0], 0});
+                var inputQueue = new Queue<BigInteger>(new BigInteger[]{parameters[0], 0});
                 while(intcoders.Any(i => !i.Done))
                 {
                     var outputValue = intcoders[activRunning++].Exec(inputQueue);
@@ -44,15 +44,15 @@ namespace dag7
         static void Star1()
         {
             string input = File.ReadAllText("input.txt");
-            var thrusterSignals = new List<int>();
-            var settings = Utils.GetPermutations(new List<int>(){0,1,2,3,4}, 5).Select(t => new List<int>(t));
+            var thrusterSignals = new List<BigInteger>();
+            var settings = Utils.GetPermutations(new List<BigInteger>(){0,1,2,3,4}, 5).Select(t => new List<BigInteger>(t));
             
             foreach (var parameters in settings)
             {
                 Intcoder intcoder = new Intcoder(input, OutputMode.OutputAndRunToEnd);
                 intcoder.Outputs.Add(0);
                 for (int i = 0; i < 5; i++)
-                    intcoder.Exec(new Queue<int>(new int[]{parameters[i], intcoder.Outputs.Last()}), true);
+                    intcoder.Exec(new Queue<BigInteger>(new BigInteger[]{parameters[i], intcoder.Outputs.Last()}), true);
                 thrusterSignals.Add(intcoder.Outputs.Last());
             }
             Console.WriteLine($"Star 1: {thrusterSignals.Max()}");
